requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
email-validator==2.2.0
psycopg[binary]==3.2.3
httpx==0.27.2
openai==1.46.0
python-dotenv==1.0.1
orjson==3.10.7
pydantic-ai-slim[ag-ui]==0.0.14
slowapi==0.1.9
redis==5.1.1
